% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.firstGroups}
\alias{.firstGroups}
\title{Return the first n groups of a grouped data.frame}
\usage{
.firstGroups(df, n = 1)
}
\arguments{
\item{df}{A grouped data.frame}

\item{n}{(Default: 1) How many groups should be returned?}
}
\description{
Useful for debugging, e.g. when trying to fill by but groups contain
multiple values.  Group order is determined by dplyr::group_rows
}
