% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.freducePartial}
\alias{.freducePartial}
\title{Create a list of partial functions and apply in sequence}
\usage{
.freducePartial(df, f, cls, ...)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{f}{Function that returns a (mutated) data.frame}

\item{cls}{character(1) Name of entry in ... for iterating.}

\item{...}{Extra arguments for f}
}
\value{
The result of applying function f sequentially to each of cls,
usually a data.frame.  Sequential application is key here, when we use
[.freducePartial()] a column in cls may not exist in df until f is
applied to the previous cls.
}
\description{
Apply a function f repeatedly to different columns of a data.frame.  Assumes
that ellipsis arguments contain (only) one vector entry and that function f
should be applied sequentially to each of these entries.  Does not work with
unquoted column names
}
\keyword{internal}
