% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{.printGroupMatch}
\alias{.printGroupMatch}
\title{Select a group from a data.frame, format into a printable string}
\usage{
.printGroupMatch(df, flt)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{flt}{An (unquoted) expression for using with dplyr::filter}
}
\value{
A character string, length 1, containing the results of printing
the first group of df that matches condition flt.
}
\description{
Get the first group of a grouped data.frame matching a condition and format
output for printing an error message.
}
\keyword{internal}
