% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stopIfColExists}
\alias{.stopIfColExists}
\title{Warn if a column already exists in a data.frame}
\usage{
.stopIfColExists(df, new_col)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{new_col}{character(n) Name of a column(s) to add to df}
}
\value{
Returns TRUE invisibly if "new_col" is not already in df.
Otherwise, raises an error and prompts the user to supply a different value.
}
\description{
Warn if a column already exists in a data.frame
}
\keyword{internal}
