% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tempColName}
\alias{.tempColName}
\title{Create a name for a temporary column in a data.frame}
\usage{
.tempColName(df, n = 1, nm = "TEMP")
}
\arguments{
\item{df}{A data.frame or tibble}

\item{n}{(default 1) How many temporary column names are needed?}

\item{nm}{(character(1), default "TEMP") Prefix for temporary column names}
}
\value{
Returns a name for a temporary column in df that isn't the name
of an existing column.
}
\description{
Returns a name that does not already exist in columns of a data.frame
}
\keyword{internal}
