% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.toString}
\alias{.toString}
\title{Wrapper for toString, doesn't convert NA to character and
only concatenates unique values}
\usage{
.toString(x)
}
\arguments{
\item{x}{Character vector for toString}
}
\value{
A character vector, length 1, where elements of x are concatenated
by ", ", but NAs are not included in the result.
}
\description{
Wrapper for toString, doesn't convert NA to character and
only concatenates unique values
}
\examples{
.toString(c("HELLO", "WORLD", NA))

# Compare results with:
toString(c("HELLO", "WORLD", NA))
}
