% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillByGroup.R
\name{fillByGroup}
\alias{fillByGroup}
\title{Fill NAs in data.frame by grouping}
\usage{
fillByGroup(
  df,
  group,
  fill,
  method = c("only_na", "all"),
  multiple = c("stop", "mode", "ignore")
)
}
\arguments{
\item{df}{A data.frame or tibble with missing (NA) values to be filled}

\item{group}{(character(n)) Names of column(s) to group by}

\item{fill}{(character(n)) Name(s) of column(s) to fill}

\item{method}{Either "only_na", of only missing entries should be filled,
or "all", if all entries should be replaced with their group mode}

\item{multiple}{(Default: "stop") How should multiple values in columns to
be filled be handled?  Either "stop" (raise an error), "mode" (select
the most common value) or "ignore" (entries with multiple possible modes are
set to NA).}
}
\description{
Wrapper of tidyr::fill with checks for NAs in grouping values and the option
to fill with the majority value if there is more than one value per group.
If an error is raised, an example group that causes an error is returned
}
