% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillTotalseq.R
\name{fillTotalseq}
\alias{fillTotalseq}
\title{Fill empty gene identifiers using TotalSeq information}
\usage{
fillTotalseq(df, ab = "Antigen")
}
\arguments{
\item{df}{A data.frame containing column(s) to fill with gene identifiers.}

\item{ab}{character(1) Name of the column in df containing antibody names.}
}
\description{
Fill empty gene identifiers using TotalSeq information
}
