% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatIg.R
\name{formatIg}
\alias{formatIg}
\title{Format immunoglobin antibody names}
\usage{
formatIg(df, ig = "Antigen", new_col = "Ig")
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ig}{(character(1), default "Antigen") Name of column containing antibody
names}

\item{new_col}{(character(1), default "Ig") Name of column containing
re-formatted immunoglobin names to add to df}
}
\value{
df, with a new column, by default named "Ig", containing the
immunoglobulin names formatted to match HGNC symbols.
}
\description{
Adds a column to a data.frame containing immunoglobin names re-formatted to
match HGNC symbols.

Assumes that Greek symbols have been converted to single letters.  Diversity
and joining regions are not implemented.  The final a/b for e.g. IgG2a is
dropped, as the specific modification isn't in protein ontology or HGNC.
The lambda light chain will only be matched in HGNC if a gene number is
provided.
}
\examples{
df <- data.frame("Antigen" = c("IgG", "IgA", "IgG1", "IgG2a"))
formatIg(df)
}
\author{
Helen Lindsay
}
