% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTCR.R
\name{formatTCR}
\alias{formatTCR}
\title{Convert T-cell receptor antigen names}
\usage{
formatTCR(df, tcr = "TCR", new_col = "TCR_long")
}
\arguments{
\item{df}{data.frame or tibble}

\item{tcr}{(character(1)) Name of column in df containing TCR names}

\item{new_col}{(character(1)) Name of column to}
}
\value{
df in long form with one subunit per row and a new column "new_col"
containing the TCR gene description query string

df, with an additional column, by default named "TCR_long",
giving the TCR name in long format to allow exact matching in the
gene_aliases table.
}
\description{
Convert T-cell antigen names to long format used in HGNC gene descriptions,
e.g. "T cell receptor gamma variable 24.  Greek symbols and words should be
converted to letters before using this function
}
\examples{
df <- data.frame(Antigen = c("TCRab", "TCR-ab", "TCRa/b"))
formatTCR(df, tcr = "Antigen")
}
\author{
Helen Lindsay
}
