% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAliases.R
\name{getAliases}
\alias{getAliases}
\title{getAliases ----}
\usage{
getAliases(ab, by = c("ALT_ID", "HGNC_ID"))
}
\arguments{
\item{ab}{(character(1)) Name of the antibody/gene to match}

\item{by}{One of ALT_ID or HGNC_ID}
}
\value{
A table of aliases, or nothing if no aliases are found
}
\description{
Given an antibody name, return aliases
}
\details{
Convenience function to return aliases all aliases in the gene_aliases table,
given a gene symbol or antibody name.  Aliases can be found by matching either
the ALT_ID (default) or HGNC_ID.  Matching by the ALT_ID means that isoforms
are not considered aliases.  For example CD45RA will not be an alias of
CD45RO.  Matching by HGNC_ID means that aliases corresponding to isoforms or
modifications of the same gene will all be returned.
}
\author{
Helen Lindsay
}
