% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonName.R
\name{getCommonName}
\alias{getCommonName}
\title{Find most common antibody name}
\usage{
getCommonName(
  x,
  cols = NULL,
  ab = "Antigen",
  fill_col = "Antigen_std",
  n_matched = "n_matched",
  ignore = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame for finding common name}

\item{cols}{Columns for grouping, default: NULL means that the columns
Antigen, Cat_Number, Clone, ALT_ID are used.  Note that gene identifiers
are not used by default as some antibodies target particular isoforms.}

\item{ab}{Column containing antibody names for standardising,
default "Antigen"}

\item{fill_col}{column to add}

\item{n_matched}{name of column to add with number of matching results,
default "n_matched"}

\item{ignore}{A named list, where names match colnames of x and values are
regular expressions matching values in x that should not be used for
grouping. For example, entries where the catalogue number
is "Custom made" should not be used for matching.  Alternatively, set to TRUE
to use the default ignore patterns.}

\item{verbose}{Print information about which columns are used for grouping?
Default: TRUE}

\item{...}{pass keep = TRUE for keeping grouping columns for debugging}
}
\description{
Find most common name by matching any/all of Antigen name, clone, gene
identifiers
}
\details{
Be careful of exceptions, e.g. Cat_Number == "custom made"
}
\author{
Helen Lindsay
}
