% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillByGroup.R
\name{groupMode}
\alias{groupMode}
\title{Find the most common value per group}
\usage{
groupMode(df, cl, gp, new_cl = NA, min_n = NA, n = NA)
}
\arguments{
\item{df}{a grouped tibble}

\item{cl}{Name of column find mode}

\item{gp}{Name(s) of columns to group by}

\item{new_cl}{Name of column to create.  If NA (default), col is modified}

\item{min_n}{(integer(1), default NA) Minimum number of occurrences of
majority value.  If provided, the majority value will be set to NA when it
occurs less than min_n times.}

\item{n}{Name of column containing counts.}
}
\description{
Given a grouped data.frame, count values per group and return a vector with
the most common value for each group.  If there are several equally common
values, the first will be chosen.
}
