% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc_long}
\alias{hgnc_long}
\title{Gene ids, names and aliases from HGNC, in long format}
\format{
A data frame with 90788 rows and 7 variables:
\describe{
    \item{HGNC_ID}{HGNC gene IDs}
    \item{ENSEMBL_ID}{Ensembl gene ID corresponding to HGNC_ID}
    \item{alias_name}{Names for gene aliases, separated by "|"}
    \item{prev_name}{Names for former gene aliases, separated by "|"}
    \item{UNIPROT_IDS}{Uniprot ID(s) corresponding to HGNC_ID}
    \item{symbol_type}{Source of the "value" column, e.g. "HGNC_SYMBOL"}
    \item{value}{A gene symbol or name}

}
}
\usage{
hgnc_long
}
\description{
Contains the same data as hgnc, but aliases, previous aliases and names have
been split to contain one entry per row, with an additional "symbol_type" column
giving the source of the symbol (e.g. HGNC_SYMBOL, alias_symbol).  Chromosome
data is not included
}
\keyword{datasets}
