% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{lowerNoDash}
\alias{lowerNoDash}
\title{Convert values to lowercase and replace - with space}
\usage{
lowerNoDash(ab)
}
\arguments{
\item{ab}{(character(n)) A vector of strings to transform}
}
\description{
Convert a character vector to lowercase and replace - with spaces.
Values are NA if the new value is identical to the original value.
}
\author{
Helen Lindsay
}
