% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchToCiteseq.R
\name{matchToCiteseq}
\alias{matchToCiteseq}
\title{Standardise names using the citeseq data set}
\usage{
matchToCiteseq(x, cols = NULL, verbose = TRUE, check.matches = FALSE, ...)
}
\arguments{
\item{x}{A data.frame or tibble containing a column "Antigen" to match to the
citeseq data set}

\item{cols}{(character(n), default NULL) Optional additional columns to use
for matching in citeseq data set.  ID columns from the citeseq dataset are
used by default.  Antibodies are grouped by a match in any of the ID columns.}

\item{verbose}{Report which columns are used for matching? (default: TRUE)}

\item{check.matches}{Set to TRUE to check which entries were matched
(default FALSE)}

\item{...}{Options passed on to getCommonName}
}
\value{
If check.matches is FALSE, the data.frame x is returned with an
additional
two columns "Antigen_std", containing the standardised name, and "n_matched",
the number of entries used for deriving the standard name.  The minimum
value of "n_matched" is 1 and means that no matches were found.

If check.matches is TRUE, a grouped tibble is returned, containing rows from x
grouped with rows from the citeseq data set that were matched.  This can be
used for checking whether matches are correct and identifying what was matched
incorrectly if they are incorrect.
}
\description{
Groups antibody names and selects the most frequent name.  By
default, the citeseq dataset is grouped by Antigen, Clone,
Cat_Number (catalog number) and ALT_ID.
}
\examples{
df <- data.frame(Antigen = c("PD-L1", "CD3"))
matchToCiteseq(df)
}
\author{
Helen Lindsay
}
