% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{print_n}
\alias{print_n}
\title{Print a data.frame n rows at a time}
\usage{
print_n(df, n = 20)
}
\arguments{
\item{df}{data.frame to print}

\item{n}{number of rows to print, default 20}
}
\value{
The first n rows of df, by default 20, and a console prompt to
either print the next n rows or quit.
}
\description{
Print a data.frame n rows at a time
}
