% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{removePrefix}
\alias{removePrefix}
\title{Remove a prefix from a column in a data.frame}
\usage{
removePrefix(df, ab = "Antigen", prefix = "[Aa]nti-", new_col = NA)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen) Name of the column to remove
prefixes from}

\item{prefix}{(character(1)) A regular expression for matching prefixes}

\item{new_col}{(character(1), default NA Name of the column to add to df.
If NA, column ab is modified}
}
\description{
Create a new column by removing a prefix from an existing
column.  The default prefix removes either "anti-" or "Anti-".  If prefix
is not found, value will be NA.
}
