% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameADT.R
\name{renameADT}
\alias{renameADT}
\alias{renameADT,SingleCellExperiment,character-method}
\alias{renameADT,MultiAssayExperiment,character-method}
\title{Replace ADT names in an object containing ADT expression}
\usage{
renameADT(obj, names, ...)

\S4method{renameADT}{SingleCellExperiment,character}(obj, names, assay = "counts", ...)

\S4method{renameADT}{MultiAssayExperiment,character}(obj, names, ..., assay = "counts")
}
\arguments{
\item{obj}{An object containing ADT expression measurements}

\item{names}{A character vector of new names, equal to the number of ADTs
in obj}

\item{...}{Not currently used}

\item{assay}{Name of the assay to be renamed (Default: "counts")}
}
\value{
An object of the same class as obj, with ADT measurements renamed
}
\description{
Given an object containing antibody derived tag (ADT) expression measurements,
such as a [SingleCellExperiment::SingleCellExperiment()] or
[MultiAssayExperiment::MultiAssayExperiment()] and a vector of new names,
replace the names of the ADTs and store the original names as metadata.

Used for e.g. standardising protein names across studies.  See
`getCommonName()` for generating standardised names by matching to the
citeseq data set. Names have to be provided because we recommend manually
checking standard names.
}
\author{
Helen Lindsay
}
