% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceGreekSyms.R
\name{replaceGreekSyms}
\alias{replaceGreekSyms}
\title{Replace Greek symbols or symbol names}
\usage{
replaceGreekSyms(x, replace = c("sym2letter", "sym2word", "word2letter"))
}
\arguments{
\item{x}{Character vector (n) containing Greek symbols to be replaced.}

\item{replace}{Either "sym2letter" (replace Greek symbols with single
lowercase letter) or "sym2word" (replace Greek symbols with lowercase word,
e.g. "alpha") or "word2letter" (replace lowercase Greek symbol names with
single lowercase letters e.g. "alpha" to "a")}
}
\value{
Character vector (n) with Greek symbols or names replaced
}
\description{
Replace Greek symbols with either single letter or word.  Note that this only
replaces greek symbols I have encountered so far in antibody names to
avoid introducing errors by e.g. replacing "mu" in "immunoglobin".
}
