% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchHGNC.R
\name{searchHGNC}
\alias{searchHGNC}
\title{searchHGNC}
\usage{
searchHGNC(df)
}
\arguments{
\item{df}{A data.frame (or similar object, e.g. a tibble)}
}
\description{
Search HGNC gene symbols, aliases and names for an exact match
to an value, assumed to be an antigen name or part of an antigen name.
}
