% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTotalseq.R
\name{searchTotalseq}
\alias{searchTotalseq}
\title{Annotate gene IDs using totalseq data set}
\usage{
searchTotalseq(x, cols = NULL)
}
\arguments{
\item{x}{data.frame, containing at minimum columns named "Antigen"}

\item{cols}{list of columns to match sequentially in totalseq.  If not
specified, defaults to catalogue number then antigen then clone.}
}
\description{
Search for exact matches between antigen names in df and totalseq cocktails
or between catalogue numbers. Relies on column names matching totalseq
As the goal is to add gene IDs, clone is not used.
}
