% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{separateBrackets}
\alias{separateBrackets}
\title{separateBrackets}
\usage{
separateBrackets(df, ab, new_col = NA)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen) Name of the column to remove
prefixes from}

\item{new_col}{(character(1), default NA Name of the column to add to df.
If NA, column ab is modified}
}
\description{
separateBrackets
}
