% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedSubstr.R
\name{sharedSubstr}
\alias{sharedSubstr}
\title{Group based on shared substrings}
\usage{
sharedSubstr(qr_df, x = "value", id = "ID", new_col = "AB_group")
}
\arguments{
\item{qr_df}{A query data.frame, e.g. created by makeQueryTable}

\item{x}{Name of column to check for shared substrings
(character(1), default: "value")}

\item{id}{Name of ID column uniquely identifying rows
(character(1), default: "ID")}

\item{new_col}{Name of column to be added to qr_df
(character(1), default: "AB_group")}
}
\description{
Takes a vector of words and a corresponding vector of group ids
and returns a numeric vector indicating groupings where members of a
group share a word with at least one other member of the group.
}
