% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitUnnest.R
\name{splitUnnest}
\alias{splitUnnest}
\title{Split a column and create one row per entry}
\usage{
splitUnnest(
  df,
  ab = "Antigen",
  split = "[\\\\(\\\\)]",
  new_col = NA,
  exclude = NA
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen) Name of the column to remove
prefixes from}

\item{split}{(character(1), default "[\\(\\)]") A regular expression for
using with strsplit. The default expression splits at "(" or ")".}

\item{new_col}{(character(1), default NA Name of the column to add to df.
If NA, column ab is modified.}

\item{exclude}{(character(1), default NA) a regex - do not split if ab
matches.}
}
\description{
Convenience function for splitting a column at a delimiter,
unnesting (one row per value after splitting) and removing unnecessary
whitespace.  Default is to split at brackets.  Returns a tibble tbl_df.
}
\examples{
df <- data.frame(Antigen = c("CD279 (PD-1)", "Mac-2 (Galectin-3)"))
splitUnnest(df, "Antigen", new_col = "Split")

}
