% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{totalseq}
\alias{totalseq}
\title{Antibody and gene information for TotalSeq antibodies from BioLegend}
\format{
A data frame with 1466 rows and 9 variables:
\describe{
    \item{Cat_Number}{BioLegend catalogue number}
    \item{Oligo_ID}{BioLegend identifier for the oligo Barcode_Sequence}
    \item{Antigen}{Name of the antigen (antibody name minus "anti-" prefix)}
    \item{Clone}{Name of the cell line that produced the antibody}
    \item{TotalSeq_Cat}{TotalSeq sequencing chemistry.  A, B, C or D.}
    \item{Reactivity}{Species in which the antibody will bind, usually Human}
    \item{Cross_Reactivity}{Species in which the antibody shows,
    cross-reactivity, where known}
    \item{Barcode_Sequence}{Sequence of DNA oligo conjuated to the antibody}
    \item{ENSEMBL_ID}{Ensembl gene identifier}
}
}
\source{
\url{"https://www.biolegend.com/"}
}
\usage{
totalseq
}
\description{
This table contains information downloaded from www.biolegend.com and
preprocessed into a consistent format.  Discontinued antibodies
do not appear in this table, and non-human gene identifiers have been
discarded.
}
\details{
Note that we do not consider the mapping between antibodies and gene
identifiers to be very reliable.  We have removed cases where the same
antibody clone was mapped to different gene identifiers, but have not
checked that the mapping agrees with the HGNC information.
}
\keyword{datasets}
