% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{totalseq_cocktails}
\alias{totalseq_cocktails}
\title{Antibody and gene information for TotalSeq cocktails from BioLegend}
\format{
A data frame with 485 rows and 8 variables:
\describe{
    \item{Oligo_ID}{BioLegend identifier for the oligo Barcode_Sequence}
    \item{Antigen}{Name of the antigen (antibody name minus "anti-" prefix)}
    \item{Clone}{Name of the cell line that produced the antibody}
    \item{Ensembl_ID}{Ensembl gene identifier}
    \item{Gene_Symbol}{Ensembl gene symbol}
    \item{TotalSeq_Cat}{Identifier for the TotalSeq cocktail}
    \item{Barcode_Sequence}{Sequence of DNA oligo conjuated to the antibody}
    \item{Reactivity}{Species in which the antibody will bind, usually human}
}
}
\source{
\url{"https://www.biolegend.com/"}
}
\usage{
totalseq_cocktails
}
\description{
This table contains information downloaded from www.biolegend.com and
preprocessed into a consistent format.  The TotalSeq A, B, C and D
(Heme Oncology) cocktails are included.
}
\keyword{datasets}
