% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{union_join}
\alias{union_join}
\title{Select rows matching any column in another data.frame}
\usage{
union_join(df, df2 = NULL, rows = NULL, by = NULL)
}
\arguments{
\item{df}{A data.frame from which to select matching rows}

\item{df2}{Optional, a second data}

\item{rows}{Row indices for subsetting, either df2 if present or df}

\item{by}{= columns to select from df2}
}
\description{
Select values from a data.frame df matching any
column from another data.frame or a selection of row indices.  If a second
data frame and a selection of rows is provided, values from df matching
any value in df2[rows, ] are returned.  If only rows indices are provided,
rows matching any value in df[rows, ] are returned.  NAs are not matched.
}
\keyword{internal}
