% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{addID}
\alias{addID}
\title{add ID column}
\usage{
addID(
  df,
  id_cols = c("Antigen", "Study"),
  new_col = "ID",
  warn = TRUE,
  sep = "__"
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{id_cols}{(character(n)) Names of columns to paste to form ID column}

\item{new_col}{(character(1), default: "ID") Name of new ID column}

\item{warn}{(TRUE/FALSE, default: TRUE) If TRUE, warn if IDs are not unique}

\item{sep}{(Default: __) Delimiter to use for pasting columns to form ID}
}
\value{
df with an extra ID column
}
\description{
Adds an ID column to a data frame
}
\details{
Pastes a group of columns together to form an ID column.  If pasted values do
not uniquely identify rows, adds a number to the end.
}
