% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{addID}
\alias{addID}
\title{add ID column}
\usage{
addID(
  df,
  id_cols = c("Antigen", "Study"),
  new_col = "ID",
  warn = TRUE,
  sep = "__",
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{id_cols}{(character(n)) Names of columns to paste to form ID column}

\item{new_col}{(character(1), default: "ID") Name of new ID column}

\item{warn}{(TRUE/FALSE, default: TRUE) If TRUE, warn if IDs are not unique}

\item{sep}{(Default: __) Delimiter to use for pasting columns to form ID}

\item{overwrite}{Should new_col be regenerated if it already exists in df?
(logical(1), default: TRUE)}

\item{verbose}{Should program status messages be printed?  Default: TRUE}
}
\value{
df with an extra ID column

df with an additional ID column, by default named "ID", which
uniquely identifies each row.  The ID column is created by pasting together
the columns named in id_cols, and adding numbers if this is insufficient
to uniquely identify rows.
}
\description{
Adds an ID column to a data frame
}
\details{
Pastes a group of columns together to form an ID column.  If pasted values do
not uniquely identify rows, adds a number to the end.  If ID column already
exists, this function will check if the ID uniquely identifies rows and will
do nothing if this is the case.  If the ID column exists and does not uniquely
identify rows, it will be regenerated and overwritten.
}
\examples{
df <- data.frame(Study = c("A", "B", "B"),
                Antigen = c("CD4","CD8","CD8"))
addID(df)
}
\author{
Helen Lindsay
}
