% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{dashNotDot}
\alias{dashNotDot}
\title{Convert dashes (-) to dots(.)}
\usage{
dashNotDot(ab)
}
\arguments{
\item{ab}{(character(n)) A vector of strings to transform}
}
\description{
Convert a character vector to lowercase and replace "-" with "."
Values are NA if the new value is identical to the original value.
}
\author{
Helen Lindsay
}
