% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{.getGroups}
\alias{.getGroups}
\title{Return the first n groups of a grouped data.frame}
\usage{
.getGroups(df, i = 1, n = 1, row_idxs = NULL)
}
\arguments{
\item{df}{A grouped data.frame}

\item{i}{(integer(1), default: 1) The index of the first group to return}

\item{n}{(integer(1), default: 1) How many groups should be returned?}

\item{row_idxs}{(Optional, default: NULL) Indices of rows}
}
\description{
Useful for debugging, e.g. when trying to fill by but groups contain
multiple values.  Group order is determined by dplyr::group_rows
}
\author{
Helen Lindsay
}
\keyword{internal}
