% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.printf}
\alias{.printf}
\title{Wrapper for sprintf, doesn't convert NA to character}
\usage{
.printf(pattern, ...)
}
\arguments{
\item{pattern}{A pattern to use with sprintf}

\item{...}{Extra arguments for sprintf}
}
\description{
Wrapper for sprintf, doesn't convert NA to character
}
