% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{.printGroupMatch}
\alias{.printGroupMatch}
\title{Select a group from a data.frame, format into a printable string}
\usage{
.printGroupMatch(df, flt)
}
\description{
Get the first group of a grouped data.frame matching a condition and format
output for printing an error message.
}
