% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stopIfColExists}
\alias{.stopIfColExists}
\title{Warn if a column already exists in a data.frame}
\usage{
.stopIfColExists(df, new_col)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{new_col}{character(n) Name of a column(s) to add to df}
}
\description{
Warn if a column already exists in a data.frame
}
\keyword{internal}
