% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tempColName}
\alias{.tempColName}
\title{Create a name for a temporary column in a data.frame}
\usage{
.tempColName(df)
}
\arguments{
\item{df}{A data.frame or tibble}
}
\description{
Returns a name that does not already exist in columns of a data.frame
}
