% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.toString}
\alias{.toString}
\title{Wrapper for toString, doesn't convert NA to character and
only concatenates unique values}
\usage{
.toString(x)
}
\arguments{
\item{x}{Character vector for toString}
}
\description{
Wrapper for toString, doesn't convert NA to character and
only concatenates unique values
}
