% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.warnIfColExists}
\alias{.warnIfColExists}
\title{Warn if a column already exists in a data.frame}
\usage{
.warnIfColExists(df, new_col)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{new_col}{character(1) Name of a column to add to df}
}
\description{
Warn if a column already exists in a data.frame
}
