% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillByGroup.R
\name{fillByGroup}
\alias{fillByGroup}
\title{Fill NAs in data.frame by grouping}
\usage{
fillByGroup(
  df,
  group,
  fill,
  method = c("only_na", "all"),
  multiple = c("stop", "mode", "ignore")
)
}
\arguments{
\item{df}{A data.frame or tibble with missing (NA) values to be filled}

\item{group}{(character(n)) Names of column(s) to group by}

\item{fill}{(character(n)) Name(s) of column(s) to fill}

\item{method}{Either "only_na", of only missing entries should be filled,
or "all", if all entries should be replaced with their group mode}

\item{multiple}{(Default: "stop") How should multiple values in columns to
be filled be handled?  Either "stop" (raise an error), "mode" (select
the most common value) or "ignore" (entries with multiple possible modes are
set to NA).}
}
\value{
df, with NA entries filled where possible by grouping and taking
the most common value within each column and group.
}
\description{
Wrapper of tidyr::fill with checks for NAs in grouping values and the option
to fill with the majority value if there is more than one value per group.
If an error is raised, an example group that causes an error is returned
}
\examples{

df <- data.frame(A = rep(c("A", "B"), each = 3),
                B = c(NA, "C", "D", "E", NA, "E"))
# Setting multiple = "mode" means that the most common value will be
# used for filling, or the first if there are ties.
fillByGroup(df, group = "A", fill = "B", multiple = "mode")

# Setting multiple = "ignore" means that groups with multiple values
# will not be filled.
fillByGroup(df, group = "A", fill = "B", multiple = "ignore")
}
\author{
Helen Lindsay
}
