% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillByGroup.R
\name{fillByGroup}
\alias{fillByGroup}
\title{Fill NAs in data.frame by grouping}
\usage{
fillByGroup(df, groups, fill, multiple = c("stop", "majority"))
}
\arguments{
\item{df}{A data.frame or tibble with missing (NA) values to be filled}

\item{groups}{(character(n)) Names of column(s) to group by}

\item{fill}{(character(n)) Name(s) of column(s) to fill}

\item{multiple}{(Default: "stop") How should multiple values in columns to
be filled be handled?  Either "stop" (raise an error) or "majority" (select
the most common value)}
}
\description{
Wrapper of tidyr::fill with checks for NAs in grouping values and the option
to fill with the majority value if there is more than one value per group.
If rlang is available and an error is raised, an example group that causes an
error is returned
}
