% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTCR.R
\name{formatTCRv}
\alias{formatTCRv}
\title{Format TCR (vector)}
\usage{
formatTCRv(tcr, new_col = "TCR_long")
}
\description{
Used by formatTCR, accepts a vector (one column from a data.frame),
returns a long data.frame with a new column containing the TCR description
}
