% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillByGroup.R
\name{groupMode}
\alias{groupMode}
\title{Find the most common value per group}
\usage{
groupMode(
  df,
  cl,
  gp,
  new_cl = NA,
  min_n = NA,
  n = NA,
  keep_first = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{df}{a grouped tibble}

\item{cl}{Name of column to find mode}

\item{gp}{Name(s) of columns to group by}

\item{new_cl}{Name of column to create.  If NA (default), col is modified}

\item{min_n}{(integer(1), default NA) Minimum number of occurrences of
majority value.  If provided, the majority value will be set to NA when it
occurs less than min_n times.}

\item{n}{Name of column containing counts.}

\item{keep_first}{(logical(1), default TRUE) If there are multiple modes,
should the first (in order of the data.frame) be selected?  If FALSE,
entries are set to NA}

\item{overwrite}{(logical(1), default FALSE) If FALSE, new_cl will only
replace NAs.  If TRUE, new_cl will contain the mode.  Note that
keep_first = FALSE with overwrite = TRUE may lead to values being replaced
with NA.}
}
\description{
Given a grouped data.frame, count values per group and return a vector with
the most common value for each group.  If there are several equally common
values, the first will be chosen.
}
