% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{gsubAb}
\alias{gsubAb}
\title{Convenience function to remove a pattern from a column in a data.frame}
\usage{
gsubAb(
  df,
  ab = "Antigen",
  pattern = "[Aa]nti-([Hh]uman)?([ _]?)",
  replacement = "",
  new_col = NA
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen) Name of the column to remove
prefixes from}

\item{pattern}{(character(1)) A regular expression for matching in column ab.}

\item{replacement}{(character(1)) Replacement value, default "" (i.e. remove)}

\item{new_col}{(character(1), default NA Name of the column to add to df.
If NA, column ab is modified}
}
\description{
Remove a pattern from a column, either modifying in place or creating a new
column.  The default pattern removes the prefix "anti-" or "Anti-", optionally
followed by Human or human, then optionally a space or underscore
}
