% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc}
\alias{hgnc}
\title{Gene names, symbols and IDs from HGNC}
\format{
A data frame with 30123 rows and 10 variables:
\describe{
    \item{HGNC_ID}{HGNC gene IDs}
    \item{HGNC_SYMBOL}{HGNC gene symbol}
    \item{HGNC_NAME}{HGNC gene name (short description in words)}
    \item{alias_symbol}{comma separated list of current gene symbol aliases}
    \item{prev_symbol}{comma separated list of formerly used symbol aliases}
    \item{ENSEMBL_ID}{Ensembl gene ID corresponding to HGNC_ID}
    \item{Chromosome}{Chromosomal location of the gene}
    \item{alias_name}{Names for gene aliases, separated by "|"}
    \item{prev_name}{Names for former gene aliases, separated by "|"}
    \item{UNIPROT_IDS}{Uniprot ID(s) corresponding to HGNC_ID}
}
}
\source{
\url{"https://www.genenames.org/cgi-bin/genegroup/download-all"}
}
\usage{
hgnc
}
\description{
A table of gene ids, symbols, aliases, previous aliases, and names
from the Human Genome Naming Consortium, corresponding to genome build
GRCh38.
}
\keyword{datasets}
