% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{makeQueryTable}
\alias{makeQueryTable}
\title{Separate Antibody/Antigen names into component parts}
\usage{
makeQueryTable(df, ab = "Antigen", control_col = NA)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default: "Antigen") Name of column in df containing
antibody names}

\item{control_col}{(character(1), default: NA)  Optional name of a logical
column indicating whether an antibody is an isotype control.  If present,
controls will be removed to avoid spurious matches as they usually do not
react against human genes.}
}
\description{
Given a data.frame containing Antigen/Antibody names, reformat
the names into possible gene names and return the data.frame in long format.
}
