% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{makeQueryTable}
\alias{makeQueryTable}
\title{Separate Antibody/Antigen names into component parts}
\usage{
makeQueryTable(df, ab = "Antigen", id_cols = c("Antigen", "Study"), funs)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen") Name of the column in df
containing Antigen/Antibody names}

\item{id_cols}{(character(n)) Names of columns to paste to form ID column}

\item{funs}{A vector of string formatting functions to}
}
\description{
Given a data.frame containing Antigen/Antibody names, reformat
the names into possible gene names and return the data.frame in long format.
}
