% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameADT.R
\name{matchToCiteseq}
\alias{matchToCiteseq}
\title{Standardise names using the citeseq data set}
\usage{
matchToCiteseq(x, cols = NULL)
}
\arguments{
\item{x}{A data.frame or tibble containing a column "Antigen" to match to the
citeseq data set}

\item{cols}{(character(n), default NULL) Optional additional columns to use
for matching in citeseq data set.  ID columns from the citeseq dataset are
used by default.  Antibodies are grouped by a match in any of the ID columns.}
}
\description{
Groups antibody names and selects the most frequent name.  By
default, the citeseq dataset is grouped by Antigen, Clone, Cat_Number
(catalog number), HGNC_ID and ENSEMBL_ID.
}
