% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPerGroup.R
\name{nPerGroup}
\alias{nPerGroup}
\title{Group a data.frame and count unique values}
\usage{
nPerGroup(df, group, col)
}
\arguments{
\item{df}{a data.frame or tibble}

\item{group}{character(n) Character vector of name(s) of grouping columns}

\item{col}{character(n) Name of column(s) in which to count distinct values}
}
\description{
Grouping columns containing NAs are not considered.
}
