% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchHGNC.R
\name{searchHGNC}
\alias{searchHGNC}
\title{searchHGNC}
\usage{
searchHGNC(query_df)
}
\arguments{
\item{query_df}{A data.frame in long format for searching for exact matches to
HGNC names.  Must contain a column "value" containing potential gene symbols
for matching with HGNC symbols and descriptions and a column "ID" for grouping
results}
}
\description{
Search HGNC gene symbols, aliases and names for an exact match
to an value, assumed to be an antigen name or part of an antigen name.
}
