% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTotalseq.R
\name{searchTotalseq}
\alias{searchTotalseq}
\title{Annotate gene IDs using totalseq data set}
\usage{
searchTotalseq(x, cols = NULL)
}
\arguments{
\item{x}{data.frame, containing at minimum columns named "Antigen"}

\item{cols}{list of columns to match in TotalSeq.  If not
specified, defaults to catalogue number, antigen and clone.}
}
\description{
Search for exact matches between antigen names or catalogue numbers in
data.frame x and totalseq cocktails data set.  Relies on column names in
matching column names in totalseq.
This function works by repeated left_joins and can add rows.  Only rows
that do not already have identifiers are joined with the totalseq data.
}
