% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{separateSubunits}
\alias{separateSubunits}
\title{Separate multi-subunit protein names}
\usage{
separateSubunits(df, ab)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{ab}{(character(1), default "Antigen) Name of the column containing
antibody names}
}
\description{
Separate names of antibodies against multi-subunit proteins
e.g. CD235ab, CD66ace into one subunit per rows,.  Assumes that a
multi-subunit protein contains at least 2 capital letters or numbers followed
by at least 2 lower case letters, where there may be a separator between the
groups and/or between the lower case letters.

At present, the between group separators are -, . and space, and the between
subunit separator is /.  Subunits should be converted from Greek symbols
before applying this function.
}
