% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedSubstr.R
\name{sharedSubstr}
\alias{sharedSubstr}
\title{Group based on shared substrings}
\usage{
sharedSubstr(x, id)
}
\arguments{
\item{x}{Character vector of words}

\item{id}{Vector of group ids}
}
\description{
Takes a vector of words and a corresponding vector of group ids
and returns a numeric vector indicating groupings where members of a
group share a word with at least one other member of the group.
}
\examples{
# "fox" occurs in groups 1 and 3, and "box" occurs in groups 3 and 4,
# so these groups are grouped together
sharedSubstr(c("fox", "cat", "fox", "in", "box", "box"), c(1, 2, 3, 3, 3, 4))
}
