% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{showGroups}
\alias{showGroups}
\title{Interactively print individual groups of a grouped data.frame}
\usage{
showGroups(df, i = 1, max_rows = 50)
}
\arguments{
\item{df}{A *grouped* data.frame}

\item{i}{integer(1) Index of the group to be printed (Default: 1).}

\item{max_rows}{integer(1) Maximum number of rows to print (Default: 50).}
}
\description{
Print first group of a grouped data.frame, then prompt user to either print
the next group or quit.   Group order is determined by dplyr::group_rows.
}
