% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGroups.R
\name{showGroups}
\alias{showGroups}
\title{Interactively print individual groups of a grouped data.frame}
\usage{
showGroups(df, i = 1, n = 1, max_rows = 50, interactive = TRUE)
}
\arguments{
\item{df}{A *grouped* data.frame}

\item{i}{(integer(1), default 1) Index of the group to be printed.}

\item{n}{(integer(1), default 1) How many groups to show at once?}

\item{max_rows}{integer(1) Maximum number of rows to print (Default: 50).}

\item{interactive}{(default: TRUE) Should the function wait for a
command prompt to show the next group?}
}
\description{
Print first group of a grouped data.frame, then prompt user to either print
the next group or quit.   Group order is determined by dplyr::group_rows.
}
