% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryTable.R
\name{upperSquish}
\alias{upperSquish}
\title{Convert values to uppercase and remove punctuation}
\usage{
upperSquish(ab)
}
\arguments{
\item{ab}{(character(n)) A vector of strings to transform}
}
\description{
Convert a character vector to uppercase and remove dashes, spaces and dots.
Values are NA if the new value is identical to the original value and are
only converted if there the original value has 2 segments separated by,
punctuation, e.g. "IFN-g" would become "IFNG" but "IFN-g R alpha-chain" would
have new value NA.  Punctuation is not removed if it separates groups of
numbers, e.g. "CD3.1"
}
\author{
Helen Lindsay
}
